+++ src/main/java/com/example/cardapio/dto/FoodRegister.java
@@ -1,4 +1,4 @@
 package com.example.cardapio.dto;
 
-public record FoodRegister(String title, String image, Integer price, String description) {
+public record FoodRegister(String title, String image, Float price, String description) {
 }
